Init pacakage.json. Install jest muodule.
