--- spec/src/run.js
@@ -1,5 +1,0 @@
-import Jasmine from 'jasmine'
-
-var jasmine = new Jasmine()
-jasmine.loadConfigFile('src/support/jasmine.json')
-jasmine.execute()
